+++ src/idpconfgen/libs/libcalc.py
@@ -272,7 +272,6 @@ def calc_torsion_angles(
     return -ARCTAN2(sin_theta, cos_theta)
 
 
-
 def calc_MSMV(data):
     """Calculate Mean, STD, Median, and Variance."""
     return np.mean(data), np.std(data), np.median(data), np.var(data)
@@ -870,121 +869,6 @@ def round_radian_to_degree_bin_10(x0):
 def unit_vector(vector):
     """Calculate the unitary vector."""
     return vector / np.linalgn.norm(vector)
-
-
-def init_lennard_jones_calculator(acoeff, bcoeff):
-    """
-    Calculate Lennard-Jones full pontential.
-
-    The LJ potential is calculated fully and no approximations to
-    proximity of infinite distance are considered.
-
-    Parameters
-    ----------
-    acoeff, bcoeff : np.ndarray, shape (N, 3), dtype=np.float
-        The LJ coefficients prepared already for the ij-pairs upon which
-        the resulting function is expected to operate.
-        IMPORTANT: it is up to the user to define the coefficients such
-        that resulting energy is np.nan for non-relevant ij-pairs, for
-        example, covalently bonded pairs, or pairs 2 bonds apart.
-
-    Returns
-    -------
-    numba.njitted func
-        Function closure with registered `acoeff`s and `bcoeff`s that
-        expects an np.ndarray of distances with same shape as `acoeff`
-        and `bcoeff`: (N,).
-        `func` returns an integer.
-    """
-    @njit
-    def calc_lennard_jones(distances_ij, NANSUM=np.nansum):
-        ar = acoeff / (distances_ij ** 12)
-        br = bcoeff / (distances_ij ** 6)
-        energy_ij = ar - br
-        return NANSUM(energy_ij)
-    return calc_lennard_jones
-
-
-def init_coulomb_calculator(charges_ij):
-    """
-    Calculate Coulomb portential.
-
-    Parameters
-    ----------
-    charges_ij : np.ndarray, shape (N, 3), dtype=np.float
-        The `charges_ij` prepared already for the ij-pairs upon which
-        the resulting function is expected to operate.
-        IMPORTANT: it is up to the user to define the charge such
-        that resulting energy is np.nan for non-relevant ij-pairs, for
-        example, covalently bonded pairs, or pairs 2 bonds apart.
-
-    Returns
-    -------
-    numba.njitted func
-        Function closure with registered `charges_ij` that expects an
-        np.ndarray of distances with same shape as `acoeff` and `bcoeff`:
-        (N,).
-        `func` returns an integer.
-    """
-    @njit
-    def calculate(distances_ij, NANSUM=np.nansum):
-        return NANSUM(charges_ij / distances_ij)
-    return calculate
-
-
-def energycalculator_ij(distf, efuncs):
-    """
-    Calculate the sum of energy terms.
-
-    This function works as a closure.
-
-    Accepts only energy terms that compute for non-redundant ij-pairs.
-
-    Energy terms must have distances ij as unique positional parameter,
-    and should return an integer.
-
-    Example
-    -------
-    >>> ecalc = energycalculator_ij(calc_ij_pair_distances, [...])
-    >>> total_energy = ecalc(coords)
-
-    Where `[...]` is a list containing energy term functions.
-
-    See Also
-    --------
-    init_lennard_jones_calculator
-    init_coulomb_calculator
-
-    Parameters
-    ----------
-    distf : func
-        The function that will be used to calculate ij-pair distances
-        on each call. If performance is a must, this function should be
-        fast. `distf` function should receive `coords` as unique
-        argument where `coords` is a np.ndarray of shape (N, 3), where N
-        is the number of atoms, and 3 represents the XYZ coordinates.
-        This function should return a np.ndarray of shape
-        (N * (N - 1)) / 2,), dtype=np.float.
-
-    efuncs : list
-        A list containing the energy terms functions. Energy term
-        functions are prepared closures that accept the output of
-        `distf` function.
-
-    Returns
-    -------
-    func
-        A function that accepts coords in the form of (N, 3). The
-        coordinates sent to the resulting function MUST be aligned with
-        the labels used to prepare the `efuncs` closures.
-    """
-    def calculate(coords):
-        dist_ij = distf(coords)
-        energy = 0
-        for func in efuncs:
-            energy += func(dist_ij)
-        return energy
-    return calculate
 
 
 # njit